# Implement Liquid::ResourceLimits