--- a/ext/liquid_c/vm.h
+++ b/ext/liquid_c/vm.h
@@ -3,7 +3,6 @@
 
 #include <ruby.h>
 #include "block.h"
-#include "resource_limits.h"
 
 enum opcode {
     OP_LEAVE = 0,