… ext/liquid_c/vm.c
… lib/liquid/c.rb
… test/unit/resource_limits_test.rb