Implements `Liquid::ResourceLimits` in C. It is roughly a translation of the Ruby code into C. I was not able to measure a significant speedup on the benchmarks.