@peterzhu2118 2020-10-01
Thanks for the feedback @dylanahsmith! I changed the benchmark files to use capture tags, I think there is a slight improvement in performance, however the results are pretty noisy on my computer.

Master:
```
              parse:    129.494  (±10.0%) i/s -      1.287k in  10.046882s
             render:    135.550  (± 8.9%) i/s -      1.356k in  10.093187s
     parse & render:     62.487  (± 6.4%) i/s -    624.000  in  10.033199s
```

This branch:
```
              parse:    131.418  (± 6.8%) i/s -      1.313k in  10.041110s
             render:    149.923  (± 8.7%) i/s -      1.498k in  10.075368s
     parse & render:     65.270  (± 6.1%) i/s -    654.000  in  10.067960s
```